Initial spike of an agent for serverdensity
Unfortunately, the sd plugins system may not work nicely with the
frozen package system of bosh. Ultimately, sd plugins system is not good
as you need to go onto each server to install plugin, or distribute it
some other way.

This bosh release will need to include all plugins that want to be supported
and allow the per-plugin token to be specified in the deployment manifest